Fix token-based user account views to have their logic in POST
This is to work around a reported bug where some user's email
or chat apps will preview links before the user can click them.
Previewing the link makes a GET request and would therefore
invalidate the single-use token, making it impossible for the
user to sign up or log in.