Work around bug TinyMCE bug preventing review submission
When a review field that uses the TinyMCE widget is
marked as required, the browser will attempt HTML
validation on form submit but will fail because
the associated <textarea required> is not focusable
(because it has display: none).

The workaround is to add "novalidate" on the whole
form, which was already the case for the review edit
form.